Final Update
TODO1: Bug Testing

+++ meteorite/src/meteorite/Pos_info.java
@@ -4,15 +4,15 @@ package meteorite;
  * Created by Alan on 21/1/2015.
  */
 public class Pos_info {
-    public int x;
-    public int y;
+    public long x;
+    public long y;
 
     public Pos_info() {
         x = 0;
         y = 0;
     }
 
-    public Pos_info(int x, int y) {
+    public Pos_info(long x, long y) {
         this.x = x;
         this.y = y;
     }
