+++ meteorite/src/meteorite/Wrapper_info.java
@@ -4,15 +4,15 @@ package meteorite;
  * Created by Alan on 21/1/2015.
  */
 public class Wrapper_info {
-    public int width;
-    public int height;
+    public long width;
+    public long height;
 
     public Wrapper_info() {
         width = 0;
         height = 0;
     }
 
-    public Wrapper_info(int w, int h) {
+    public Wrapper_info(long w, long h) {
         this.width = w;
         this.height = h;
     }
